Refactor folder structure to use the hero module
Remove traces of active record usages.
