+++ src/hero/hero-class/models/hero-class.ts
@@ -1,11 +1,11 @@
-import {BaseEntity, Column, Entity, OneToMany, PrimaryGeneratedColumn} from "typeorm";
+import {Column, Entity, OneToMany, PrimaryGeneratedColumn} from "typeorm";
 import {Field, ID, ObjectType} from "type-graphql";
 import * as jf from "joiful";
 import {HeroAscendancy} from "../../hero-ascendancy/models/hero-ascendancy";
 
 @Entity("hero_class")
 @ObjectType()
-class HeroClass extends BaseEntity {
+class HeroClass {
     @Field(() => ID)
     @jf.number().positive().required()
     @PrimaryGeneratedColumn()
