+++ src/hero/hero-class/resolvers/hero-class.ts
